+++ Makefile
@@ -6,7 +6,7 @@
 #    By: tafujise <tafujise@student.42.jp>          +#+  +:+       +#+         #
 #                                                 +#+#+#+#+#+   +#+            #
 #    Created: 2025/10/30 12:42:52 by tafujise          #+#    #+#              #
-#    Updated: 2026/02/15 02:36:50 by tafujise         ###   ########.fr        #
+#    Updated: 2026/02/16 02:54:01 by tafujise         ###   ########.fr        #
 #                                                                              #
 # **************************************************************************** #
 
@@ -38,6 +38,7 @@ SRCS:=main.c\
 	  execute/redirect/stdio_guard.c\
 	  execute/utils/args_utils.c\
 	  execute/utils/path_utils.c\
+	  execute/utils/envp_utils.c\
 	  hashmap/hashmap_crud.c\
 	  hashmap/hashmap_free.c\
 	  hashmap/hashmap_utils.c
