+++ src/execute/process/wait_children.c
@@ -1,12 +1,12 @@
 /* ************************************************************************** */
 /*                                                                            */
 /*                                                        :::      ::::::::   */
-/*   collect_child.c                                    :+:      :+:    :+:   */
+/*   wait_children.c                                    :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
 /*   By: tafujise <tafujise@student.42.jp>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2026/01/27 23:52:55 by tafujise          #+#    #+#             */
-/*   Updated: 2026/02/11 11:03:22 by tafujise         ###   ########.fr       */
+/*   Updated: 2026/02/16 02:16:39 by tafujise         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -15,17 +15,17 @@
 int			status_to_exitcode(int status);
 void		reset_ctx_pid(t_ctx *ctx);
 
+/*
+	Todo
+	- Wait the arrays of pids
+	- Convert last pid's status to exit code and store in ctx->exit_code
+	- Clean-up ctx's the arrays of pids.
+*/
 t_status	collect_child_result(t_ctx *ctx)
 {
 	int	i;
 	int	status;
 
-	/*
-		Todo
-		- Wait the arrays of pids
-		- Convert last pid's status to exit code and store in ctx->exit_code
-		- Clean-up ctx's the arrays of pids.
-	*/
 	i = 0;
 	status = 0;
 	if (ctx->pids == NULL || ctx->npid < 1)
