+++ test/unit_test/disk/Makefile
@@ -6,7 +6,7 @@
 #    By: tafujise <tafujise@student.42.jp>          +#+  +:+       +#+         #
 #                                                 +#+#+#+#+#+   +#+            #
 #    Created: 2026/02/15 15:31:59 by tafujise          #+#    #+#              #
-#    Updated: 2026/02/15 15:37:17 by tafujise         ###   ########.fr        #
+#    Updated: 2026/02/16 02:53:31 by tafujise         ###   ########.fr        #
 #                                                                              #
 # **************************************************************************** #
 
@@ -34,6 +34,7 @@ EXECUTE_SRCS := \
 	exec_disk.c\
 	args_utils.c\
 	path_utils.c\
+	envp_utils.c\
 
 HASHMAP_SRCS := $(notdir $(wildcard $(ROOT)/src/hashmap/*.c))
